reduce duplication in search results trees
