+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/classifications-tree/internal/visibility/SearchResultsTree.ts
@@ -6,15 +6,21 @@
 import { firstValueFrom } from "rxjs";
 import { assert } from "@itwin/core-bentley";
 import { CLASS_NAME_Classification, CLASS_NAME_ClassificationTable } from "../../../common/internal/ClassNameDefinitions.js";
-import { getOrCreate, ParentElementsPath } from "../../../common/internal/Utils.js";
-import { createSearchResultsTree, SearchResultsNodesHandler } from "../../../common/internal/visibility/BaseSearchResultsTree.js";
+import { ParentElementsPath } from "../../../common/internal/Utils.js";
+import {
+  addElementToInternalSearchTargets,
+  convertInternalSearchTargetElements,
+  createSearchResultsTree,
+  SearchResultsNodesHandler,
+} from "../../../common/internal/visibility/BaseSearchResultsTree.js";
 
 import type { Id64Set, Id64String } from "@itwin/core-bentley";
 import type { HierarchySearchTree } from "@itwin/presentation-hierarchies";
 import type { EC, ECClassHierarchyInspector } from "@itwin/presentation-shared";
-import type { CategoryId, ElementId, ModelId } from "../../../common/internal/Types.js";
+import type { ElementId } from "../../../common/internal/Types.js";
 import type {
   BaseSearchResultsTreeNode,
+  InternalSearchTargetElements,
   SearchResultsTree,
   SearchResultsTreeNodeChildren,
   SearchResultsTreeRootNode,
@@ -80,17 +86,6 @@ type RawElementNode = Omit<ElementNode, "modelId" | "categoryId" | "children" | 
 };
 
 type RawNode = RawClassificationTableNode | RawClassificationNode | RawElementNode;
-
-type InternalSearchTargetElements = Map<
-  ModelId,
-  Map<
-    ElementId | undefined,
-    {
-      parentElementsPath: ParentElementsPath;
-      elements: Map<CategoryId, { searchTargets: Array<ElementId>; nonSearchTargets: Array<ElementId> }>;
-    }
-  >
->;
 
 interface InternalSearchTargets {
   elements?: InternalSearchTargetElements;
@@ -162,26 +157,6 @@ class ClassificationsTreeSearchResultsNodesHandler extends SearchResultsNodesHan
     return this.convertInternalSearchTargets(internalSearchTargets);
   }
 
-  private convertInternalSearchTargetElements(
-    internalSearchTargetElements: InternalSearchTargetElements,
-  ): Required<ClassificationsTreeSearchTargets>["elements"] {
-    const result: Required<ClassificationsTreeSearchTargets>["elements"] = [];
-    for (const [modelId, modelEntry] of internalSearchTargetElements) {
-      for (const { parentElementsPath, elements } of modelEntry.values()) {
-        for (const [categoryId, { searchTargets, nonSearchTargets }] of elements) {
-          result.push({
-            categoryId,
-            modelId,
-            parentElementsPath,
-            nonSearchTargetElements: nonSearchTargets,
-            searchTargetElements: searchTargets,
-          });
-        }
-      }
-    }
-    return result;
-  }
-
   private convertInternalSearchTargets(searchTargets: InternalSearchTargets): ClassificationsTreeSearchTargets | undefined {
     if (!searchTargets.classificationTableIds && !searchTargets.classificationIds && !searchTargets.elements) {
       return undefined;
@@ -190,7 +165,7 @@ class ClassificationsTreeSearchResultsNodesHandler extends SearchResultsNodesHan
     return {
       classificationTableIds: searchTargets.classificationTableIds,
       classificationIds: searchTargets.classificationIds,
-      elements: searchTargets.elements ? this.convertInternalSearchTargetElements(searchTargets.elements) : undefined,
+      elements: searchTargets.elements ? convertInternalSearchTargetElements(searchTargets.elements) : undefined,
     };
   }
 
@@ -228,39 +203,7 @@ class ClassificationsTreeSearchResultsNodesHandler extends SearchResultsNodesHan
         // Internal search target elements need to have path saved in some way.
         // For this, a tree structure is used, where keys are stringified identifiers of parent nodes depending on the hierarchy.
         internalSearchTargets.elements ??= new Map();
-        const modelEntry = getOrCreate({
-          map: internalSearchTargets.elements,
-          key: node.modelId,
-          createFunc: () =>
-            new Map<
-              ElementId | undefined,
-              {
-                parentElementsPath: ParentElementsPath;
-                elements: Map<CategoryId, { searchTargets: Array<ElementId>; nonSearchTargets: Array<ElementId> }>;
-              }
-            >(),
-        });
-        const lastParentId = ParentElementsPath.getSingleLastParentId(node.parentElementsPath);
-        const parentEntry = getOrCreate({
-          map: modelEntry,
-          key: lastParentId,
-          createFunc: () => ({
-            parentElementsPath: node.parentElementsPath,
-            elements: new Map<
-              CategoryId,
-              {
-                searchTargets: Array<ElementId>;
-                nonSearchTargets: Array<ElementId>;
-              }
-            >(),
-          }),
-        });
-        const categoryEntry = getOrCreate({ map: parentEntry.elements, key: node.categoryId, createFunc: () => ({ searchTargets: [], nonSearchTargets: [] }) });
-        if (node.isSearchTarget) {
-          categoryEntry.searchTargets.push(node.id);
-        } else {
-          categoryEntry.nonSearchTargets.push(node.id);
-        }
+        addElementToInternalSearchTargets(internalSearchTargets.elements, node);
       }
     }
   }
