+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/models-tree/internal/visibility/SearchResultsTree.ts
@@ -7,7 +7,12 @@ import { firstValueFrom } from "rxjs";
 import { assert } from "@itwin/core-bentley";
 import { CLASS_NAME_Category, CLASS_NAME_Model, CLASS_NAME_Subject } from "../../../common/internal/ClassNameDefinitions.js";
 import { getOrCreate, ParentElementsPath } from "../../../common/internal/Utils.js";
-import { createSearchResultsTree, SearchResultsNodesHandler } from "../../../common/internal/visibility/BaseSearchResultsTree.js";
+import {
+  addElementToInternalSearchTargets,
+  convertInternalSearchTargetElements,
+  createSearchResultsTree,
+  SearchResultsNodesHandler,
+} from "../../../common/internal/visibility/BaseSearchResultsTree.js";
 
 import type { Id64Set, Id64String } from "@itwin/core-bentley";
 import type { HierarchySearchTree } from "@itwin/presentation-hierarchies";
@@ -15,6 +20,7 @@ import type { EC, ECClassHierarchyInspector } from "@itwin/presentation-shared";
 import type { CategoryId, ElementId, ModelId } from "../../../common/internal/Types.js";
 import type {
   BaseSearchResultsTreeNode,
+  InternalSearchTargetElements,
   SearchResultsTree,
   SearchResultsTreeNodeChildren,
   SearchResultsTreeRootNode,
@@ -90,16 +96,6 @@ type RawElementNode = Omit<ElementNode, "children" | "modelId" | "parentElements
 
 type RawNode = RawElementNode | RawModelNode | RawSubjectNode | RawCategoryNode;
 
-type InternalSearchTargetElements = Map<
-  ModelId,
-  Map<
-    ElementId | undefined,
-    {
-      parentElementsPath: ParentElementsPath;
-      elements: Map<CategoryId, { searchTargets: Array<ElementId>; nonSearchTargets: Array<ElementId> }>;
-    }
-  >
->;
 type InternalSearchTargetCategories = Map<
   ModelId,
   Map<
@@ -217,25 +213,6 @@ class ModelsTreeSearchResultsNodesHandler extends SearchResultsNodesHandler<Proc
     return this.convertInternalSearchTargets(internalSearchTargets);
   }
 
-  private convertInternalSearchTargetElements(internalSearchTargetElements: InternalSearchTargetElements): Required<ModelsTreeSearchTargets>["elements"] {
-    const result: Required<ModelsTreeSearchTargets>["elements"] = [];
-    // Internal search target elements are stored in a tree structure, need to convert that to array structure.
-    for (const [modelId, modelEntry] of internalSearchTargetElements) {
-      for (const { parentElementsPath, elements } of modelEntry.values()) {
-        for (const [categoryId, { searchTargets, nonSearchTargets }] of elements) {
-          result.push({
-            categoryId,
-            modelId,
-            parentElementsPath,
-            nonSearchTargetElements: nonSearchTargets,
-            searchTargetElements: searchTargets,
-          });
-        }
-      }
-    }
-    return result;
-  }
-
   private convertInternalSearchTargetCategories(
     internalSearchTargetCategories: InternalSearchTargetCategories,
   ): Required<ModelsTreeSearchTargets>["categories"] {
@@ -260,7 +237,7 @@ class ModelsTreeSearchResultsNodesHandler extends SearchResultsNodesHandler<Proc
 
     return {
       categories: searchTargets.categories ? this.convertInternalSearchTargetCategories(searchTargets.categories) : undefined,
-      elements: searchTargets.elements ? this.convertInternalSearchTargetElements(searchTargets.elements) : undefined,
+      elements: searchTargets.elements ? convertInternalSearchTargetElements(searchTargets.elements) : undefined,
       modelIds: searchTargets.modelIds,
       subjectIds: searchTargets.subjectIds,
     };
@@ -321,39 +298,7 @@ class ModelsTreeSearchResultsNodesHandler extends SearchResultsNodesHandler<Proc
       }
       case "element": {
         internalSearchTargets.elements ??= new Map();
-        const modelEntry = getOrCreate({
-          map: internalSearchTargets.elements,
-          key: node.modelId,
-          createFunc: () =>
-            new Map<
-              ElementId | undefined,
-              {
-                parentElementsPath: ParentElementsPath;
-                elements: Map<CategoryId, { searchTargets: Array<ElementId>; nonSearchTargets: Array<ElementId> }>;
-              }
-            >(),
-        });
-        const lastParentId = ParentElementsPath.getSingleLastParentId(node.parentElementsPath);
-        const parentEntry = getOrCreate({
-          map: modelEntry,
-          key: lastParentId,
-          createFunc: () => ({
-            parentElementsPath: node.parentElementsPath,
-            elements: new Map<
-              CategoryId,
-              {
-                searchTargets: Array<ElementId>;
-                nonSearchTargets: Array<ElementId>;
-              }
-            >(),
-          }),
-        });
-        const categoryEntry = getOrCreate({ map: parentEntry.elements, key: node.categoryId, createFunc: () => ({ searchTargets: [], nonSearchTargets: [] }) });
-        if (node.isSearchTarget) {
-          categoryEntry.searchTargets.push(node.id);
-        } else {
-          categoryEntry.nonSearchTargets.push(node.id);
-        }
+        addElementToInternalSearchTargets(internalSearchTargets.elements, node);
       }
     }
   }
